Moved structures to seperate headers
