+++ Proj/Project_1/P1_The_War_Game/Setup.h
@@ -1,0 +1,22 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+/* 
+ * File:   Setup.h
+ * Author: Kevin Maltz
+ *
+ * Created on April 21, 2018, 4:45 PM
+ */
+
+#ifndef SETUP_H
+#define SETUP_H
+
+namespace setup{
+    
+}
+
+#endif /* SETUP_H */
+
