+++ Proj/Project_1/P1_The_War_Game/nbproject/project.xml
@@ -6,7 +6,7 @@
             <name>P1_The_War_Game</name>
             <c-extensions/>
             <cpp-extensions>cpp</cpp-extensions>
-            <header-extensions/>
+            <header-extensions>h</header-extensions>
             <sourceEncoding>UTF-8</sourceEncoding>
             <make-dep-projects/>
             <sourceRootList/>
